chore: fix up tests
