+++ packages/ipfs-core-utils/src/files/normalise-input/content.js
@@ -2,7 +2,12 @@
 
 const errCode = require('err-code')
 const { Buffer } = require('buffer')
-const { Blob, FileReader } = require('ipfs-utils/src/globalthis')
+const {
+  isBytes,
+  isBloby,
+  browserStreamToIt,
+  blobToIt
+} = require('./utils')
 
 function toAsyncIterable (input) {
   // Bytes | String
@@ -14,7 +19,7 @@ function toAsyncIterable (input) {
 
   // Bloby
   if (isBloby(input)) {
-    return blobToAsyncGenerator(input)
+    return blobToIt(input)
   }
 
   // Browser stream
@@ -65,68 +70,6 @@ function toAsyncIterable (input) {
 
 function toBuffer (chunk) {
   return isBytes(chunk) ? chunk : Buffer.from(chunk)
-}
-
-function blobToAsyncGenerator (blob) {
-  if (typeof blob.stream === 'function') {
-    // firefox < 69 does not support blob.stream()
-    return browserStreamToIt(blob.stream())
-  }
-
-  return readBlob(blob)
-}
-
-async function * readBlob (blob, options) {
-  options = options || {}
-
-  const reader = new FileReader()
-  const chunkSize = options.chunkSize || 1024 * 1024
-  let offset = options.offset || 0
-
-  const getNextChunk = () => new Promise((resolve, reject) => {
-    reader.onloadend = e => {
-      const data = e.target.result
-      resolve(data.byteLength === 0 ? null : data)
-    }
-    reader.onerror = reject
-
-    const end = offset + chunkSize
-    const slice = blob.slice(offset, end)
-    reader.readAsArrayBuffer(slice)
-    offset = end
-  })
-
-  while (true) {
-    const data = await getNextChunk()
-
-    if (data == null) {
-      return
-    }
-
-    yield Buffer.from(data)
-  }
-}
-
-function isBytes (obj) {
-  return Buffer.isBuffer(obj) || ArrayBuffer.isView(obj) || obj instanceof ArrayBuffer
-}
-
-function isBloby (obj) {
-  return obj instanceof Blob
-}
-
-async function * browserStreamToIt (stream) {
-  const reader = stream.getReader()
-
-  while (true) {
-    const result = await reader.read()
-
-    if (result.done) {
-      return
-    }
-
-    yield result.value
-  }
 }
 
 module.exports = toAsyncIterable
